requirements.txt
boto3==1.34.44
botocore==1.34.44
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
einops==0.7.0
encodec==0.1.1
filelock==3.9.0
fsspec==2024.2.0
funcy==2.0
huggingface-hub==0.20.3
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pycparser==2.21
python-dateutil==2.8.2
pytorch-triton-rocm==2.2.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.1
s3transfer==0.10.0
safetensors==0.4.2
scipy==1.12.0
six==1.16.0
soundfile==0.12.1
suno-bark @ git+https://github.com/suno-ai/bark.git@773624d26db84278a55aacae9a16d7b25fbccab8
sympy==1.12
tokenizers==0.15.2
torch==2.2.0+rocm5.7
torchaudio==2.2.0+rocm5.7
torchvision==0.17.0+rocm5.7
tqdm==4.66.2
transformers==4.37.2
typing_extensions==4.8.0
urllib3==1.26.13
